+++ software/getMatchesByDate/src/main/java/com/mtjworldcup/getbydate/model/MatchDto.java
@@ -1,6 +1,7 @@
 package com.mtjworldcup.getbydate.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public record MatchDto(String matchId, LocalDate date, LocalTime startTime, String homeTeam, String awayTeam, Integer pool) {}
+public record MatchDto(String matchId, LocalDate date, LocalTime startTime, String homeTeam, String awayTeam, BigDecimal pool) {}
